+++ test/functional/test_framework/test_framework.py
@@ -126,6 +126,7 @@ class BitcoinTestFramework(metaclass=BitcoinTestMetaClass):
         # addrman will not result in automatic connections to them.
         self.disable_autoconnect = True
         self.set_test_params()
+        assert self.wallet_names is None or len(self.wallet_names) <= self.num_nodes
         if self.options.timeout_factor == 0:
             self.options.timeout_factor = 99999
         # optionally, increase timeout by a factor
@@ -528,13 +529,19 @@ class BitcoinTestFramework(metaclass=BitcoinTestMetaClass):
                 assert_equal(chain_info["initialblockdownload"], False)
 
     def import_deterministic_coinbase_privkeys(self):
-        wallet_names = (
-            [self.default_wallet_name] * len(self.nodes)
+        for i in range(self.num_nodes):
+            self.init_wallet(i)
+
+    def init_wallet(self, i):
+        wallet_name = (
+            self.default_wallet_name
             if self.wallet_names is None
-            else self.wallet_names
+            else self.wallet_names[i]
+            if i < len(self.wallet_names)
+            else False
         )
-        assert len(wallet_names) <= len(self.nodes)
-        for wallet_name, n in zip(wallet_names, self.nodes):
+        if wallet_name is not False:
+            n = self.nodes[i]
             if wallet_name is not None:
                 n.createwallet(
                     wallet_name=wallet_name,