+++ test/functional/wallet_backup.py
@@ -87,10 +87,10 @@ class WalletBackupTest(BitcoinTestFramework):
         self.generate(self.nodes[3], 1)
 
     # As above, this mirrors the original bash test.
-    def start_three(self):
-        self.start_node(0)
-        self.start_node(1)
-        self.start_node(2)
+    def start_three(self, args=()):
+        self.start_node(0, self.extra_args[0] + list(args))
+        self.start_node(1, self.extra_args[1] + list(args))
+        self.start_node(2, self.extra_args[2] + list(args))
         self.connect_nodes(0, 3)
         self.connect_nodes(1, 3)
         self.connect_nodes(2, 3)
@@ -155,6 +155,11 @@ class WalletBackupTest(BitcoinTestFramework):
             wallet_name,
             wallet_file,
         )
+
+    def init_three(self):
+        self.init_wallet(0)
+        self.init_wallet(1)
+        self.init_wallet(2)
 
     def run_test(self):
         self.log.info("Generating initial blockchain")
@@ -231,7 +236,8 @@ class WalletBackupTest(BitcoinTestFramework):
         shutil.rmtree(os.path.join(self.nodes[2].datadir, self.chain, "blocks"))
         shutil.rmtree(os.path.join(self.nodes[2].datadir, self.chain, "chainstate"))
 
-        self.start_three()
+        self.start_three(["-nowallet"])
+        self.init_three()
 
         assert_equal(self.nodes[0].getbalance(), 0)
         assert_equal(self.nodes[1].getbalance(), 0)