+++ test/functional/wallet_multiwallet.py
@@ -45,6 +45,7 @@ class MultiWalletTest(BitcoinTestFramework):
         self.setup_clean_chain = True
         self.num_nodes = 2
         self.rpc_timeout = 120
+        self.extra_args = [["-nowallet"], []]
 
     def skip_test_if_missing_module(self):
         self.skip_if_no_wallet()
@@ -98,7 +99,9 @@ class MultiWalletTest(BitcoinTestFramework):
         # rename wallet.dat to make sure plain wallet file paths (as opposed to
         # directory paths) can be loaded
         # create another dummy wallet for use in testing backups later
-        self.start_node(0, ["-nowallet", "-wallet=empty", "-wallet=plain"])
+        self.start_node(0)
+        node.createwallet("empty", descriptors=False)
+        node.createwallet("plain", descriptors=False)
         node.createwallet("created")
         self.stop_nodes()
         empty_wallet = os.path.join(self.options.tmpdir, "empty.dat")
@@ -133,8 +136,11 @@ class MultiWalletTest(BitcoinTestFramework):
         ]
         if os.name == "nt":
             wallet_names.remove("w7_symlink")
-        extra_args = ["-nowallet"] + [f"-wallet={n}" for n in wallet_names]
-        self.start_node(0, extra_args)
+        self.start_node(0)
+        for wallet_name in wallet_names[:-2]:
+            self.nodes[0].createwallet(wallet_name, descriptors=False)
+        for wallet_name in wallet_names[-2:]:
+            self.nodes[0].loadwallet(wallet_name)
         assert_equal(
             sorted(w["name"] for w in self.nodes[0].listwalletdir()["wallets"]),
             [
@@ -152,18 +158,19 @@ class MultiWalletTest(BitcoinTestFramework):
 
         assert_equal(set(node.listwallets()), set(wallet_names))
 
+        # should raise rpc error if wallet path can't be created
+        assert_raises_rpc_error(
+            -1,
+            "filesystem error:" if sys.platform != "win32" else "create_directories:",
+            self.nodes[0].createwallet,
+            "w8/bad",
+            descriptors=False,
+        )
+
         # check that all requested wallets were created
         self.stop_node(0)
         for wallet_name in wallet_names:
             assert_equal(os.path.isfile(wallet_file(wallet_name)), True)
-
-        # should not initialize if wallet path can't be created
-        exp_stderr = (
-            "filesystem error:" if sys.platform != "win32" else "create_directories:"
-        )
-        self.nodes[0].assert_start_raises_init_error(
-            ["-wallet=w8/bad"], exp_stderr, match=ErrorMatch.PARTIAL_REGEX
-        )
 
         self.nodes[0].assert_start_raises_init_error(
             ["-walletdir=wallets"],
@@ -219,7 +226,9 @@ class MultiWalletTest(BitcoinTestFramework):
         # if wallets/ doesn't exist, datadir should be the default wallet dir
         wallet_dir2 = data_dir("walletdir")
         os.rename(wallet_dir(), wallet_dir2)
-        self.start_node(0, ["-nowallet", "-wallet=w4", "-wallet=w5"])
+        self.start_node(0)
+        self.nodes[0].createwallet("w4")
+        self.nodes[0].createwallet("w5")
         assert_equal(set(node.listwallets()), {"w4", "w5"})
         w5 = wallet("w5")
         self.generatetoaddress(
@@ -229,9 +238,9 @@ class MultiWalletTest(BitcoinTestFramework):
         # now if wallets/ exists again, but the rootdir is specified as the
         # walletdir, w4 and w5 should still be loaded
         os.rename(wallet_dir2, wallet_dir())
-        self.restart_node(
-            0, ["-nowallet", "-wallet=w4", "-wallet=w5", f"-walletdir={data_dir()}"]
-        )
+        self.restart_node(0, ["-nowallet", f"-walletdir={data_dir()}"])
+        self.nodes[0].loadwallet("w4")
+        self.nodes[0].loadwallet("w5")
         assert_equal(set(node.listwallets()), {"w4", "w5"})
         w5 = wallet("w5")
         w5_info = w5.getwalletinfo()
@@ -239,7 +248,8 @@ class MultiWalletTest(BitcoinTestFramework):
 
         competing_wallet_dir = os.path.join(self.options.tmpdir, "competing_walletdir")
         os.mkdir(competing_wallet_dir)
-        self.restart_node(0, [f"-walletdir={competing_wallet_dir}"])
+        self.restart_node(0, ["-nowallet", f"-walletdir={competing_wallet_dir}"])
+        self.nodes[0].createwallet(self.default_wallet_name, descriptors=False)
         exp_stderr = (
             r"Error: Error initializing wallet database environment"
             r" \"\S+competing_walletdir\S*\"!"
@@ -250,7 +260,9 @@ class MultiWalletTest(BitcoinTestFramework):
             match=ErrorMatch.PARTIAL_REGEX,
         )
 
-        self.restart_node(0, extra_args)
+        self.restart_node(0)
+        for wallet_name in wallet_names:
+            self.nodes[0].loadwallet(wallet_name)
 
         assert_equal(
             sorted(w["name"] for w in self.nodes[0].listwalletdir()["wallets"]),